Finished up Mod 26
Finished moving the API to rot.rs and added a decode and encode method,
which are basically just frontends for the underlying rotate function.

I also added some more tests to work on this functionality.

Finally, I added a commandline parser which just ignores the first input
and then prints out the before and after decoding for every input after.